socialDeepScopeCollector/requirements.txt
django >=4.2
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
django-filter
django-utils-six
requests
mysqlclient